Re-order database lookup
I'm not sure why this was moved; the change seems unnecessary.
Further, the change meant that the lookup actually now happens _before_
the debugging line that announces it.

So by the time the code says "Fetching enrolled students for {course}...",
the lookup has already been performed.

Unless there's reason to do so, let's leave it in place like this.